+++ src/app/locales/en/components/content-atm-inventories-workflows.js
@@ -16,5 +16,5 @@ export default {
   editorView,
   loadingView,
   taskDetailsView,
-  confirmPageClose: 'You have some unsaved changes. Your changes will be lost if you don\'t save them.',
+  confirmPageClose: 'You have unsaved changes. Your changes will be lost if you don\'t save them.',
 };