+++ CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
 ### Latest changes
 
+* VFS-8263 Showing "unsaved changes" warning when leaving workflows editor
 * VFS-8076 Rendering workflow upload input in DOM
 * VFS-7896 Added possibility to define audit log stores in workflows
 * VFS-7947 Improved URL routing to running new workflow with initial input stores values